Added evolution via repoze.evolution for: users, books, events, groups, messages, cache.
The DB will auto-evolve each sub-DB to version 1 to add the
field "evolved" which is now used by base.mak to show the
Sub-DB versions in the footer (to be removed later).
Please remember to update the evolved field in any future
evolve scripts so it can be used as a debugging tool later
on.

*** Be careful committing changes to object structures! ***
Since the version numbers must be bumped and evolution scripts
have to be written, parallel works on the same sub-DB must be
carefully coordonated. Simple git merges will end in chaos
(= rebuild of a completely new DB every time and
   uselessness of evolution).